requirements.txt
aiogram
asyncio
logging
requests
pyYaml
